added texturing + inefficient but working draft of ppm image loader

+++ GraphicsGameEnigne/Drawable.cpp
@@ -14,7 +14,7 @@ void Drawable::draw()
 	}
 	VScbuf.bind();
 	// window draw? gfx draw?
-	gfx.pDeviceContext->DrawIndexed(this->indicesCount(), 0, 0);
+	gfx.pDeviceContext->DrawIndexed(static_cast<UINT>(this->indicesCount()), 0, 0);
 }
 
 void Drawable::addBind(std::unique_ptr<Bindable>&& b)
